Extract shared connection presentation helpers (DT-4174)
Both ComputeBadge (list pill) and ConnectionBadge (details column) duplicated
the state->color map, state->label, and tooltip-text logic. Move them to
connection-status.ts (connectionStateColor, connectionStateLabel,
connectionTooltip) so there is a single source of truth; the components are
now thin presentational shells.
