+++ src/lib/components/deployments/compute-badge.svelte
@@ -1,13 +1,12 @@
 <script lang="ts">
-  import { cva } from 'class-variance-authority';
-
   import Icon from '$lib/holocene/icon/icon.svelte';
   import Tooltip from '$lib/holocene/tooltip.svelte';
-  import { translate } from '$lib/i18n/translate';
   import type { ComputeStatus } from '$lib/types/deployments';
   import {
+    connectionStateColor,
+    connectionStateLabel,
+    connectionTooltip,
     deriveConnectionStatus,
-    formatConnectionCheckTime,
   } from '$lib/utilities/connection-status';
 
   const CONFIG: Record<string, { icon: 'aws' | 'gcp'; label: string }> = {
@@ -24,40 +23,6 @@
   const state = $derived(
     computeStatus ? deriveConnectionStatus(computeStatus) : undefined,
   );
-
-  const connectionText = cva([], {
-    variants: {
-      state: {
-        connected: 'text-success',
-        failed: 'text-danger',
-        pending: 'text-subtle',
-      },
-    },
-  });
-
-  const connectionLabel = $derived.by((): string => {
-    if (state === 'connected')
-      return translate('deployments.connection-connected');
-    if (state === 'failed') return translate('deployments.connection-failed');
-    return translate('deployments.connection-pending');
-  });
-
-  const tooltipText = $derived.by((): string => {
-    if (state === 'pending') {
-      return translate('deployments.connection-tooltip-pending');
-    }
-    const time = formatConnectionCheckTime(
-      computeStatus?.providerValidation?.lastCheckTime,
-    );
-    if (state === 'connected') {
-      return translate('deployments.connection-tooltip-connected', { time });
-    }
-    const errorMessage = computeStatus?.providerValidation?.errorMessage ?? '';
-    return (
-      (errorMessage ? `${errorMessage}. ` : '') +
-      translate('deployments.connection-tooltip-failed-checked', { time })
-    );
-  });
 </script>
 
 {#snippet pill()}
@@ -68,18 +33,25 @@
     <p>{config.label}</p>
     {#if state}
       <span
-        class="size-1.5 shrink-0 rounded-full bg-current {connectionText({
-          state,
-        })}"
+        class="size-1.5 shrink-0 rounded-full bg-current {connectionStateColor[
+          state
+        ]}"
       ></span>
-      <span class={connectionText({ state })}>{connectionLabel}</span>
+      <span class={connectionStateColor[state]}>
+        {connectionStateLabel(state)}
+      </span>
     {/if}
   </div>
 {/snippet}
 
 {#if config}
   {#if state}
-    <Tooltip text={tooltipText} topLeft width={250} usePortal>
+    <Tooltip
+      text={connectionTooltip(computeStatus)}
+      topLeft
+      width={250}
+      usePortal
+    >
       {@render pill()}
     </Tooltip>
   {:else}
