+++ src/lib/utilities/connection-status.ts
@@ -1,5 +1,6 @@
 import { formatDistanceToNowStrict } from 'date-fns';
 
+import { translate } from '$lib/i18n/translate';
 import type { ComputeStatus } from '$lib/types/deployments';
 import type { ValidTime } from '$lib/utilities/format-time';
 import { isTimestamp, timestampToDate } from '$lib/utilities/format-time';
@@ -12,6 +13,19 @@ export const deriveConnectionStatus = (
   const validation = computeStatus?.providerValidation;
   if (!validation?.lastCheckTime) return 'pending';
   return validation.errorMessage ? 'failed' : 'connected';
+};
+
+export const connectionStateColor: Record<ConnectionState, string> = {
+  connected: 'text-success',
+  failed: 'text-danger',
+  pending: 'text-subtle',
+};
+
+export const connectionStateLabel = (state: ConnectionState): string => {
+  if (state === 'connected')
+    return translate('deployments.connection-connected');
+  if (state === 'failed') return translate('deployments.connection-failed');
+  return translate('deployments.connection-pending');
 };
 
 export const formatConnectionCheckTime = (time: ValidTime): string => {
@@ -29,4 +43,22 @@ export const formatConnectionCheckTime = (time: ValidTime): string => {
   } catch {
     return 'less than an hour ago';
   }
+};
+
+export const connectionTooltip = (computeStatus?: ComputeStatus): string => {
+  const state = deriveConnectionStatus(computeStatus);
+  if (state === 'pending') {
+    return translate('deployments.connection-tooltip-pending');
+  }
+  const time = formatConnectionCheckTime(
+    computeStatus?.providerValidation?.lastCheckTime,
+  );
+  if (state === 'connected') {
+    return translate('deployments.connection-tooltip-connected', { time });
+  }
+  const errorMessage = computeStatus?.providerValidation?.errorMessage ?? '';
+  return (
+    (errorMessage ? `${errorMessage}. ` : '') +
+    translate('deployments.connection-tooltip-failed-checked', { time })
+  );
 };
